path: rename from graph/path to query/path
