+++ query/graphql/graphql.go
@@ -15,8 +15,8 @@ import (
 	"github.com/dennwc/graphql/language/parser"
 
 	"github.com/cayleygraph/cayley/graph"
-	"github.com/cayleygraph/cayley/graph/path"
 	"github.com/cayleygraph/cayley/query"
+	"github.com/cayleygraph/cayley/query/path"
 	"github.com/cayleygraph/quad"
 )
