+++ query/path/morphism_apply_functions.go
