+++ query/path/pathtest/pathtest.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 	"time"
 
-	. "github.com/cayleygraph/cayley/graph/path"
+	. "github.com/cayleygraph/cayley/query/path"
 
 	"github.com/cayleygraph/cayley/graph"
 	"github.com/cayleygraph/cayley/graph/graphtest/testutil"
