+++ schema/loader.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/cayleygraph/cayley/graph/path"
+	"github.com/cayleygraph/cayley/query/path"
 	"github.com/cayleygraph/quad"
 
 	"github.com/cayleygraph/cayley/graph"
